requirements.txt
Django==4.2.2
django-cryptography==1.1
django-dbbackup==4.0.2
djangorestframework==3.14.0
hdwallet==2.2.1
pytest==7.4.0
pytest-django==4.5.2
bitcoinlib==0.6.10
